+++ spring-boot/src/main/kotlin/fi/hsl/jore4/mapmatching/service/common/response/LinkTraversalDTO.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.mapmatching.service.routing.response
+package fi.hsl.jore4.mapmatching.service.common.response
 
 import fi.hsl.jore4.mapmatching.model.ExternalLinkReference
 import fi.hsl.jore4.mapmatching.model.PathTraversal
