+++ spring-boot/src/main/kotlin/fi/hsl/jore4/mapmatching/service/common/response/RoutingResponse.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.mapmatching.service.routing.response
+package fi.hsl.jore4.mapmatching.service.common.response
 
 interface RoutingResponse {
     val code: ResponseCode
