+++ spring-boot/src/main/kotlin/fi/hsl/jore4/mapmatching/service/common/response/RoutingSuccessDTO.kt
@@ -1,3 +1,3 @@
-package fi.hsl.jore4.mapmatching.service.routing.response
+package fi.hsl.jore4.mapmatching.service.common.response
 
 data class RoutingSuccessDTO(override val code: ResponseCode, val routes: List<RouteResultDTO>) : RoutingResponse
