+++ spring-boot/src/main/kotlin/fi/hsl/jore4/mapmatching/service/routing/RoutingServiceImpl.kt
@@ -7,13 +7,13 @@ import fi.hsl.jore4.mapmatching.repository.infrastructure.ILinkRepository
 import fi.hsl.jore4.mapmatching.repository.infrastructure.SnapToLinkDTO
 import fi.hsl.jore4.mapmatching.repository.routing.IRoutingRepository
 import fi.hsl.jore4.mapmatching.repository.routing.RouteLinkDTO
+import fi.hsl.jore4.mapmatching.service.common.response.RoutingFailureDTO
+import fi.hsl.jore4.mapmatching.service.common.response.RoutingResponse
+import fi.hsl.jore4.mapmatching.service.common.response.RoutingResultTransformer
+import fi.hsl.jore4.mapmatching.service.node.INodeServiceInternal
+import fi.hsl.jore4.mapmatching.service.node.NodeResolutionParams
 import fi.hsl.jore4.mapmatching.service.routing.RoutingServiceHelper.createNodeResolutionParams
 import fi.hsl.jore4.mapmatching.service.routing.RoutingServiceHelper.findUnmatchedCoordinates
-import fi.hsl.jore4.mapmatching.service.routing.internal.INodeServiceInternal
-import fi.hsl.jore4.mapmatching.service.routing.internal.NodeResolutionParams
-import fi.hsl.jore4.mapmatching.service.routing.response.RoutingFailureDTO
-import fi.hsl.jore4.mapmatching.service.routing.response.RoutingResponse
-import fi.hsl.jore4.mapmatching.service.routing.response.RoutingResultTransformer
 import fi.hsl.jore4.mapmatching.util.CollectionUtils.filterOutConsecutiveDuplicates
 import fi.hsl.jore4.mapmatching.util.LogUtils.joinToLogString
 import org.slf4j.Logger
